move tests for discontiguous indices to their own test suite
