+++ cpp/tests/opendnp3/src/TestOutstation.cpp
@@ -516,133 +516,6 @@ TEST_CASE(SUITE("ReadByRangeHeader"))
 	REQUIRE(t.lower->PopWriteAsHex() == "C2 81 80 00 1E 02 00 05 06 01 2A 00 01 29 00");
 }
 
-TEST_CASE(SUITE("ContiguousIndexesInDiscontiguousModeIntegrityScan"))
-{
-	// this will tell the outstation to use discontiguous index mode, but we won't change the address assignments
-	OutstationConfig config;
-	config.params.indexMode = IndexMode::Discontiguous;
-	OutstationTestObject t(config, DatabaseSizes::BinaryOnly(2));
-	t.LowerLayerUp();
-
-	t.SendToOutstation(hex::IntegrityPoll(0));
-
-	REQUIRE(t.lower->PopWriteAsHex() == "C0 81 80 00 01 02 00 00 01 02 02");
-}
-
-TEST_CASE(SUITE("can safely update non-existent index in discontiguous index mode"))
-{
-	OutstationConfig config;
-	config.params.indexMode = IndexMode::Discontiguous;
-	OutstationTestObject t(config, DatabaseSizes::BinaryOnly(2));
-
-	auto view = t.context.GetConfigView();
-	for (int i = 0; i < 2; ++i)
-	{
-		view.binaries[i].config.clazz = PointClass::Class1;
-		view.binaries[i].config.svariation = StaticBinaryVariation::Group1Var2;
-		view.binaries[i].config.evariation = EventBinaryVariation::Group2Var3;
-		view.binaries[i].config.vIndex = i + 2;
-	}
-
-	// now send an update to boundary points that doesn't exist
-	REQUIRE_FALSE(t.context.GetUpdateHandler().Update(Binary(true), 0, EventMode::Suppress));
-	REQUIRE_FALSE(t.context.GetUpdateHandler().Update(Binary(true), 1, EventMode::Suppress));
-	// 2 & 3 exist
-	REQUIRE_FALSE(t.context.GetUpdateHandler().Update(Binary(true), 4, EventMode::Suppress));
-}
-
-TEST_CASE(SUITE("ContiguousIndexesInDiscontiguousModeRangeScan"))
-{
-	// this will tell the outstation to use discontiguous index mode, but we won't change the address assignments
-	OutstationConfig config;
-	config.params.indexMode = IndexMode::Discontiguous;
-	OutstationTestObject t(config, DatabaseSizes::BinaryOnly(2));
-	t.LowerLayerUp();
-
-	t.SendToOutstation("C0 01 01 02 00 00 01");
-
-	REQUIRE(t.lower->PopWriteAsHex() == "C0 81 80 00 01 02 00 00 01 02 02");
-}
-
-std::string QueryDiscontiguousBinary(const std::string& request)
-{
-	OutstationConfig config;
-	config.params.indexMode = IndexMode::Discontiguous;
-
-	OutstationTestObject t(config, DatabaseSizes::BinaryOnly(3));
-
-	// assign virtual indices to the database specified above
-	auto view = t.context.GetConfigView();
-	view.binaries[0].config.vIndex = 2;
-	view.binaries[1].config.vIndex = 4;
-	view.binaries[2].config.vIndex = 5;
-
-	t.LowerLayerUp();
-
-	t.Transaction([](IUpdateHandler & db)
-	{
-		db.Update(Binary(true, 0x01), 2, EventMode::Suppress);
-		db.Update(Binary(false, 0x01), 4, EventMode::Suppress);
-	});
-
-	t.SendToOutstation(request);
-	return t.lower->PopWriteAsHex();
-}
-
-TEST_CASE(SUITE("ReadDiscontiguousClass0"))
-{
-	REQUIRE(QueryDiscontiguousBinary("C0 01 3C 01 06") == "C0 81 80 00 01 02 00 02 02 81 01 02 00 04 05 01 02");
-}
-
-TEST_CASE(SUITE("ReadDiscontiguousBadRangeBelow"))
-{
-	// read 01 var 2, [00 : 01]
-	REQUIRE(QueryDiscontiguousBinary("C0 01 01 02 00 00 01") == "C0 81 80 04");
-}
-
-TEST_CASE(SUITE("ReadDiscontiguousBadRangeAbove"))
-{
-	// read 01 var 2, [06 : 09]
-	REQUIRE(QueryDiscontiguousBinary("C0 01 01 02 00 06 09") == "C0 81 80 04");
-}
-
-
-TEST_CASE(SUITE("ReadDiscontiguousSingleRange"))
-{
-	// read 01 var 2, [02 : 02]
-	REQUIRE(QueryDiscontiguousBinary("C0 01 01 02 00 02 02") == "C0 81 80 00 01 02 00 02 02 81");
-}
-
-TEST_CASE(SUITE("ReadDiscontiguousDoubleRange"))
-{
-	// read 01 var 2, [04 : 05]
-	REQUIRE(QueryDiscontiguousBinary("C0 01 01 02 00 04 05") == "C0 81 80 00 01 02 00 04 05 01 02");
-}
-
-TEST_CASE(SUITE("ReadDiscontiguousPastUpperBound"))
-{
-	// read 01 var 2, [05 : 06]
-	REQUIRE(QueryDiscontiguousBinary("C0 01 01 02 00 05 06") == "C0 81 80 04 01 02 00 05 05 02");
-}
-
-TEST_CASE(SUITE("ReadDiscontiguousAllDataWithMultipleRanges"))
-{
-	// read 01 var 2, [02 : 02]; read 01 var 2, [04 : 05]
-	REQUIRE(QueryDiscontiguousBinary("C0 01 01 02 00 02 02 01 02 00 04 05") == "C0 81 80 00 01 02 00 02 02 81 01 02 00 04 05 01 02");
-}
-
-TEST_CASE(SUITE("ReadDiscontiguousAllDataWithMultipleRangesAndRangeError"))
-{
-	// read 01 var 2, [02 : 03]; read 01 var 2, [04 : 05]
-	REQUIRE(QueryDiscontiguousBinary("C0 01 01 02 00 02 03 01 02 00 04 05") == "C0 81 80 04 01 02 00 02 02 81 01 02 00 04 05 01 02");
-}
-
-TEST_CASE(SUITE("ReadDiscontiguousAllDataWithRangeError"))
-{
-	// read 01 var 2, [02 : 05]
-	REQUIRE(QueryDiscontiguousBinary("C0 01 01 02 00 02 05") == "C0 81 80 04 01 02 00 02 02 81 01 02 00 04 05 01 02");
-}
-
 template <class PointType>
 void TestStaticType(const OutstationConfig& config, const DatabaseSizes& tmp, PointType value, const std::string& rsp, const std::function<void (DatabaseConfigView&)>& configure)
 {
